// src/frontend/src/backend.d.ts
export interface Product {
    id: string;
    stockQuantity: bigint;
    isOnSale: boolean;
    name: string;
    description: string;
    imageUrl: string;
    category: string;
    salePrice?: bigint;
    price: bigint;
}